Changed Capacity Metric to Double type
git-svn-id: https://svnserver.informatik.kit.edu/i43/svn/code/Palladio/Incubation/ExperimentAutomation/trunk@25317 769cc9a0-fb0c-0410-8c2f-a2305f0e6268

+++ org.palladiosimulator.experimentautomation.application/src/org/palladiosimulator/experimentautomation/application/VariationFactorTuple.java
@@ -5,18 +5,18 @@ import org.palladiosimulator.experimentautomation.experiments.Variation;
 public class VariationFactorTuple {
 
     private final Variation variation;
-    private long factor;
+    private Double factor;
 
-    public VariationFactorTuple(final Variation variation, final long factor) {
+    public VariationFactorTuple(final Variation variation, final Double factor) {
         this.variation = variation;
         this.factor = factor;
     }
 
-    public long getFactor() {
+    public Double getFactor() {
         return this.factor;
     }
 
-    public void setFactor(long factor) {
+    public void setFactor(Double factor) {
         this.factor = factor;
     }
